[ECOMMERCE-PRICE-VO][TESTS] Move component specific tests to namespaces matching components #1550
